to int

+++ db/migrate/20181114144920_create_item.rb
@@ -3,16 +3,16 @@ class CreateItem < ActiveRecord::Migration[5.2]
     create_table :items do |t|
       t.string :code, null: false, default: 'DEFAULT_CODE'
       t.string :name, null: false, default: 'DEFAULT_NAME'
-      t.float :stock, null: false, default: 0
-      t.float :buy, null: false, default: 0
-      t.float :sell, null: false, default: 0
+      t.bigint :stock, null: false, default: 0
+      t.bigint :buy, null: false, default: 0
+      t.bigint :sell, null: false, default: 0
       t.references :item_cat, null: false, foreign_key: true
       t.string :brand, null: false, defautl: "DEFAULT BRAND"
-      t.float :wholesale, null: false, default: 0
-      t.float :box, null: false, default: 0
+      t.bigint :wholesale, null: false, default: 0
+      t.bigint :box, null: false, default: 0
       t.string :image
-      t.float :buy_grocer, default: 0
-      t.float :discount, null: false, default: 0
+      t.bigint :buy_grocer, default: 0
+      t.bigint :discount, null: false, default: 0
       
       t.timestamps
     end
